Services/automation_scripts/requirements.txt
celery[redis]==5.3.4
redis==4.6.0
python-dotenv==1.0.0
requests==2.31.0
flower==2.0.1
